+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## unreleased
 - Exposed the `flutter upgrade` command
+- Added a Flutter settings menu item
+- We now install our package dependencies the first time the plugin is run
 
 ## 0.1.2
 - Added a `Send Feedback…` menu item